#12235 Add an edit form for environment samples in web app - moved EnvironmentSampleController into environmentsample package
