+++ sormas-ui/src/main/java/de/symeda/sormas/ui/ControllerProvider.java
@@ -37,8 +37,8 @@ import de.symeda.sormas.ui.immunization.ImmunizationController;
 import de.symeda.sormas.ui.person.PersonController;
 import de.symeda.sormas.ui.reports.aggregate.AggregateReportController;
 import de.symeda.sormas.ui.samples.AdditionalTestController;
-import de.symeda.sormas.ui.samples.EnvironmentSampleController;
 import de.symeda.sormas.ui.samples.PathogenTestController;
+import de.symeda.sormas.ui.samples.environmentsample.EnvironmentSampleController;
 import de.symeda.sormas.ui.samples.humansample.SampleController;
 import de.symeda.sormas.ui.sormastosormas.SormasToSormasController;
 import de.symeda.sormas.ui.statistics.StatisticsController;
